Use custom json encoder
Avoids errors with datetime obtjects not being JSON serializable
